refactor: navbar z-index and hero section overflow
I added z-index: 1 to the navbar to make sure it stays above the page
content. Hero section now has overflow-y: none, and overflow-x: clip.
Additionally, the navbar and the nav-links element has a solid white
background when the menu is open on mobile viewports, due to the bug
in chromium browsers where backdrop-filter: blur() not working with
nested elements.