feat: added case card display logic and failsafe
I added functions to toggle between the three display styles of the
case cards on the "Cases" page. Function resets class of the
"card-container" div, and adds the proper class for each display upon
clicking the respective button.

I also added "if" statements to the different functions in the script.js
file, so that page-specific functions are only executed if the relevant
div is present on the currently loaded page.