feat: added info obfuscation/reveal logic, styling
I added JavaScript code to not display my email and phone number on
first visit. The user has to click the respective boxes to reveal that
information. This is then saved to localstorage, so they do not need to
click it again when revisiting of after navigating away.
The LinkedIn and Github elements also have been changed to <a> tags and
open the correct links in a new tab.