fix: case card image sizing and adding cases
I fixed the sizing of the case card thumbnail images so that it will
fill the height on default view as the cards become narrower and I added
a max-height so the image doesn't suffocate the card content at specific
sizes.
I also began adding cases to the cases page to replace the placeholder
cards.